% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rless.R
\docType{package}
\name{rless}
\alias{rless}
\alias{rless-package}
\title{rless: css preprocessor R library}
\description{
rless is R package providing CSS preprocessor features to R users.
}
\details{
It uses LESS language, which is an CSS extension giving option to use
variables, functions or using operators while creating styles. Visit oficial
LESS website (\url{http://lesscss.org}) for more information about language
specifics. The provided LESS content is converted into CSS using V8 JavaScript engine
(\url{https://v8.dev/}).
}
\author{
\strong{Maintainer}: Jonas Vaclavek \email{jonas.vaclavek@gmail.com}

Authors:
\itemize{
  \item Jakub Kuzilek
}

}
