% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rless.R
\name{convert_folder}
\alias{convert_folder}
\title{Converts files in folder to CSS files}
\usage{
convert_folder(input_folder, output_folder = tempdir(),
  recursive = FALSE, pattern = "*.less$")
}
\arguments{
\item{input_folder}{Path to files to be converted}

\item{output_folder}{Output path where converted files should be placed
Files are placed to \code{tempdir} by default}

\item{recursive}{Boolean value used when searching for filed to be converted}

\item{pattern}{Pattern which files need to match to be converted}
}
\value{
Paths to converted CSS files
}
\description{
Converts files in folder to CSS files
}
