% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_less.R
\name{parse_less}
\alias{parse_less}
\title{Parse LESS content to CSS}
\usage{
parse_less(code)
}
\arguments{
\item{code}{text to be converted in LESS format}
}
\value{
converted text to CSS
}
\description{
Parse LESS file to CSS and return it as result
}
\details{
Currently `@import` functionality is not supported.
}
\examples{
parse_less('.button { .red{ collor: red}}')
parse_less('@red: red; .red{ collor: @red}')
}
